requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyyaml>=6.0.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0